+++ source/electricity-market/ElectricityMarket.ImportOrchestrator/Orchestration/Activities/ImportGoldModelActivity.cs
@@ -18,9 +18,9 @@ using System.Data;
 using System.Diagnostics;
 using System.Dynamic;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution;
+using Energinet.DataHub.ElectricityMarket.Infrastructure.Helpers;
 using Energinet.DataHub.ElectricityMarket.Infrastructure.Options;
 using Energinet.DataHub.ElectricityMarket.Infrastructure.Persistence.Model;
-using Energinet.DataHub.ElectricityMarket.Infrastructure.Repositories;
 using FastMember;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Data.SqlClient;
